requirements.txt
transformers==4.40.2
torch==2.3.0
Pillow==10.3.0
matplotlib==3.8.4
timm==0.9.16
gradio==4.31.5
streamlit==1.35.0